Tried tty-command
It is awesome and has many good features, but vim still complains the
input and output are not the terminal. I tried it quickly, and did not
figure out what's up.

So, it's back to the popen3 streams.

Vim sends characters directly to the terminal to make the TUI.

How this sort of stuff works:

```
$ vim &
[1] 29957
$ jobs
[1]+  Stopped                 vim
$ fg
vim
$
```

What fg does?

	new file:   Gemfile
	new file:   Gemfile.lock
	modified:   ruby_comline.rb