OK, mix Process.spawn and Open3
Process.spawn is the right way to spawn a process: the parent process
sets the arguments, the file descriptors, and the environment. You do
not get pipes from a spawned process. You give them to it.

It works for the interactive calls then.

But spawning a process in background is still done with Open3. And it
does not make sense for vim and such programs. I need to figure out how
to do it properly. You can stop and re-activate processes, which are
connected to your terminal streams. But can you also redirect the
streams to a log? Can you open a pseudo-terminal for such processes, to
satisfy their requirement to have a terminal, then connect & disconnect
your TTY or just a stream file for logging and whatnot?

	modified:   ruby_comline.rb
	new file:   tests/spawn_tty.rb