A bunch of failed attempts to pass commands to another comline
I just forgot that the `sh` example in README is done with popen3, which
of course opens pipes for the process. Sending lines down these pipes
works. But If you launch the process interactively from a terminal, with
/dev/pts/N in the /proc/<pid>/fd/[012], then you cannot simply send
strings to the pts.

Let's separate the shell and the UI then. The shell will run with pipes.

	modified:   bin/comline
	modified:   lib/ruby_shell.rb